1007. Minimum Domino Rotations For Equal Row.
